1257 o crud completo

+++ src/br/com/kaio/cadswingfiel/domain/Pagamento.java
@@ -10,15 +10,15 @@ public class Pagamento {
 	@EmbeddedId
 	private PagamentoId id;
 	
-	private long valor;
+	private double valor;
 	private int tipo;
 
 
-	public long getValor() {
+	public double getValor() {
 		return valor;
 	}
 
-	public void setValor(long valor) {
+	public void setValor(double valor) {
 		this.valor = valor;
 	}
